tidy: Fix "multiplication implicit widening" in mythmusic.
Fix the clang-tidy "implicit widening of multiplication result"
warnings.  This check is looking for places where a multiplication
operation may have overflowed the smaller type before the result is
converted to the bigger type. This will give erroneous results.

1) Cast multiplication arguments to a larger type.

2) Convert a variable to the appropriate type.

3) Convert a couple of #define to typed static constexpr.

+++ mythplugins/mythmusic/mythmusic/bumpscope.h
@@ -20,7 +20,7 @@ public:
         {(void) action;}
 
 private:
-    static void blur_8(unsigned char *ptr, int w, int h, int bpl);
+    static void blur_8(unsigned char *ptr, int w, int h, ptrdiff_t bpl);
 
     void generate_cmap(unsigned int color);
     void generate_phongdat(void);
@@ -49,7 +49,7 @@ private:
     bool           m_movingLight    {true};
     //bool         m_diamond        {true};
 
-    int            m_bpl            {0};
+    ptrdiff_t      m_bpl            {0};
 
     std::vector<std::vector<unsigned char> > m_phongDat {};
     unsigned char *m_rgbBuf         {nullptr};
